Add a github-based commenting system.
The commenting system is based on github commit comments and is relatively modular.
The goal is to refactor it into its own repository and add it as a submodule for
jekyll-based sites. The javascript is linted, but still in beta stage since it hasn't
been tested with all possible edge-cases.